add the possibility to first take 'cities' db admin
since we now (https://github.com/CanalTP/cosmogony2cities) can have
international admins in `cities` db.

This PR adds a `ed2nav` flag `--cities_first` to first fetch the admins
in the database, before looking for the `admin` in the dataset (thus the
admins loaded by osm2ed)

This will makes the admins more similar between mimir&kraken.
It is far from being perfect as the kraken's admin will continue to be
different (and uglier :wink: ) than mimir, but the aim is to makes them
disapear anyway.

It should not break the existing integration (without the cli flag).